+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Version 1.0.0
 + 增加 :class:`werobot.reply.ImageReply`
 + 增加 :class:`werobot.reply.VoiceReply`
 + 增加 :class:`werobot.reply.VideoReply`
++ 删除 :func:`werobot.reply.create_reply`
 + 为 :class:`werobot.reply.WeChatReply` 增加 ``process_args`` 方法
 + 删去了 Reply 中过时的 flag
 + 修复 :class:`werobot.session.filestorage.FileStorage` 在 PyPy 下的兼容性问题